requirements.txt
Django==4.1
djangorestframework==3.14.0
drf_spectacular==0.24.2
environ==1.0
pyotp==2.6.0
stripe==4.1.0
